requirements.txt
langchain
langchain-community
sentence-transformers
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
transformers
torch
langchain-huggingface 
huggingface_hub 
transformers 
accelerate 
bitsandbytes
